chore(M): Clean Frontend Services
